+++ Magic/CHANGELOG.md
@@ -7,6 +7,7 @@
  - Make wand permissions consistent so armor wands should work the same as handheld wands
  - Add `validate_spells: false` option for config.yml, to turn off spell validation
  - Add `portal_warp` magic block parameter to turn nether portals into warps, use `/mblock add portal`
+ - Phase of moon requirement default format changed to parse as = instead of >=
 
 # 10.5.1